% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Fits elastic net to data using coordinate descent algorithm}
\usage{
elnet_coord(X, y, alpha, lambda, eps = 1e-06)
}
\arguments{
\item{X}{The predictor variable matrix}

\item{y}{The response variable matrix}

\item{alpha}{A value between 0 and 1 indicating weight toward L1 vs L2 shrinkage}

\item{lambda}{The weight parameter for determining significance of the shrinkage parameters}

\item{eps}{Stopping criteria}
}
\value{
The estimated coefficient vector beta
}
\description{
Fits elastic net to data using coordinate descent algorithm
}
\examples{
n = 100
alpha = 1
eps = matrix(rnorm(n, 0, 1), nrow=n)
X = mvrnorm(n, mu=rep(0, 20), Sigma = diag(rep(1, 20)))
X = scale(X)
beta = matrix(c(2,0,-2,0,1,0,-1,0,rep(0, 12)), nrow=20)
y = X\%*\%beta+eps
elnet_coord(X, y, alpha, 0.05)
}
