% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{solve_ols}
\alias{solve_ols}
\title{Solves a linear system using Gauss-Seidel or Jacobi method}
\usage{
solve_ols(A, b, v = NULL, iteration = 10000, method = "gauss_seidel",
  no_core = 1)
}
\arguments{
\item{A}{The coefficient matrix for solving x in b = Ax}

\item{b}{The outcome matrix for solving x in linear system b = Ax}

\item{v}{The correct outcome matrix for comparing relative error per iteration}

\item{iteration}{Optional Parameter for number of iterations desired for the algorithm; 10000 default}

\item{method}{Optional Parameter for method, choice of "gauss_seidel" or "jacobi"}

\item{no_core}{Optional Parameter for number of cores if using jacobi parallel}
}
\value{
A list of predicted x, relative error if given v, and time for each iteration
}
\description{
Solves a linear system using Gauss-Seidel or Jacobi method
}
\examples{
A = matrix(0, nrow = 100, ncol=100)
diag(A) = rep(3, 100)
index = seq(1, 99)
A[cbind(index+1, index)] = rep(-1, 99)
A[cbind(index, index+1)] = rep(-1, 99)
v = matrix(rep(1,100), nrow=100)
b = A \%*\% v
solve_ols(A, b, v, method="gauss_seidel")
solve_ols(A, b, v, method="jacobi")
}
